[shape-callout] address review comments (factor out geometry utils from Callout; remove `vectors` dependency; add some geometry vector utils to Autodiff)
